requirements.txt
fastapi
uvicorn[standard]
pydantic
paho-mqtt
pytest
python-dotenv
